Move `DatabaseSession` to a standalone file
